initial requirement - checkout as entry point, modularise app, load Products, add loading UI, render each product in its component
