+++ src/ui/ProductListUI.js
@@ -1,0 +1,21 @@
+import { ProductUI } from "./ProductUI";
+
+export const ProductListUI = ({
+  products,
+  addProductToCart,
+  removeProductFromCart,
+}) => {
+  return (
+    <div>
+      <h1>Health Products</h1>
+
+      <div className="card-container">
+        {products.map((product) => (
+          <div key={product.id} className="card">
+            <ProductUI product={product} />
+          </div>
+        ))}
+      </div>
+    </div>
+  );
+};
